+++ kala-collection/src/main/java/kala/collection/immutable/ImmutableSeq.java
@@ -242,7 +242,7 @@ public interface ImmutableSeq<@Covariant E> extends ImmutableCollection<E>, Seq<
     }
 
     @Override
-    default @NotNull ImmutableSeq<E> toImmutableSeq() {
+    default @NotNull ImmutableSeq<E> toSeq() {
         return this;
     }